refactor: project restructing
